# smv2/openapi.json
{
  "components": {
    "schemas": {
...
      "CourseOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failed_asset_count": {
            "default": 0,
            "title": "Failed Asset Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProgressSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "section_count": {
            "default": 0,
            "title": "Section Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "CourseOut",
        "type": "object"
      },
...
      "DiagnosticBlindCaseOut": {
        "properties": {
          "concept_description_md": {
            "title": "Concept Description Md",
            "type": "string"
          },
          "concept_id": {
            "title": "Concept Id",
            "type": "string"
          },
          "concept_label": {
            "title": "Concept Label",
            "type": "string"
          },
          "evidence_available": {
            "title": "Evidence Available",
            "type": "boolean"
          }
        },
        "required": [
          "concept_id",
          "concept_label",
          "concept_description_md",
          "evidence_available"
        ],
        "title": "DiagnosticBlindCaseOut",
        "type": "object"
      },
      "DiagnosticDisagreementReasonIn": {
        "properties": {
          "disagreement_reason": {
            "enum": [
              "model_estimate",
              "item_mapping",
              "concept_granularity",
              "insufficient_student_evidence",
              "instructor_disagreement"
            ],
            "title": "Disagreement Reason",
            "type": "string"
          }
        },
        "required": [
          "disagreement_reason"
        ],
        "title": "DiagnosticDisagreementReasonIn",
        "type": "object"
      },
      "DiagnosticJudgmentIn": {
        "properties": {
          "concept_id": {
            "title": "Concept Id",
            "type": "string"
          },
          "disagreement_reason": {
            "anyOf": [
              {
                "enum": [
                  "model_estimate",
                  "item_mapping",
                  "concept_granularity",
                  "insufficient_student_evidence",
                  "instructor_disagreement"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disagreement Reason"
          },
          "judgment": {
            "enum": [
              "insufficient",
              "not_struggling",
              "uncertain",
              "likely_struggling"
            ],
            "title": "Judgment",
            "type": "string"
          },
          "notes_md": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes Md"
          }
        },
        "required": [
          "concept_id",
          "judgment"
        ],
        "title": "DiagnosticJudgmentIn",
        "type": "object"
      },
      "DiagnosticJudgmentOut": {
        "properties": {
          "agreement": {
            "title": "Agreement",
            "type": "boolean"
          },
          "concept_id": {
            "title": "Concept Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "disagreement_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disagreement Reason"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "judgment": {
            "title": "Judgment",
            "type": "string"
          },
          "model_state": {
            "title": "Model State",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "readiness_estimate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readiness Estimate"
          },
          "requires_disagreement_reason": {
            "readOnly": true,
            "title": "Requires Disagreement Reason",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "concept_id",
          "judgment",
          "disagreement_reason",
          "model_state",
          "readiness_estimate",
          "evidence_count",
          "model_version",
          "agreement",
          "created_at",
          "requires_disagreement_reason"
        ],
        "title": "DiagnosticJudgmentOut",
        "type": "object"
      },
      "DiagnosticValidationSummaryOut": {
        "properties": {
          "agreement_count": {
            "title": "Agreement Count",
            "type": "integer"
          },
          "chance_adjusted_agreement": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chance Adjusted Agreement"
          },
          "disagreement_reasons": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Disagreement Reasons",
            "type": "object"
          },
          "disagreements_by_concept": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Disagreements By Concept",
            "type": "object"
          },
          "pending_reason_count": {
            "title": "Pending Reason Count",
            "type": "integer"
          },
          "raw_agreement": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Agreement"
          },
          "sample_size": {
            "title": "Sample Size",
            "type": "integer"
          },
          "sufficient_sample": {
            "title": "Sufficient Sample",
            "type": "boolean"
          }
        },
        "required": [
          "sample_size",
          "agreement_count",
          "raw_agreement",
          "chance_adjusted_agreement",
          "sufficient_sample",
          "pending_reason_count",
          "disagreement_reasons",
          "disagreements_by_concept"
        ],
        "title": "DiagnosticValidationSummaryOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PracticeAnsweredOut": {
        "properties": {
          "answered_at": {
            "format": "date-time",
            "title": "Answered At",
            "type": "string"
          },
          "correct": {
            "title": "Correct",
            "type": "boolean"
          },
          "correct_index": {
            "title": "Correct Index",
            "type": "integer"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "evidence_state": {
            "title": "Evidence State",
            "type": "string"
          },
          "explanation_md": {
            "title": "Explanation Md",
            "type": "string"
          },
          "readiness_estimate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readiness Estimate"
          },
          "selected_index": {
            "title": "Selected Index",
            "type": "integer"
          }
        },
        "required": [
          "selected_index",
          "correct",
          "correct_index",
          "explanation_md",
          "readiness_estimate",
          "evidence_state",
          "evidence_count",
          "answered_at"
        ],
        "title": "PracticeAnsweredOut",
        "type": "object"
      },
      "PracticeAssessmentOut": {
        "properties": {
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/PracticeQuestionOut"
            },
            "title": "Questions",
            "type": "array"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "section_id": {
            "title": "Section Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "ready",
              "generating",
              "failed",
              "not_started"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "section_id"
        ],
        "title": "PracticeAssessmentOut",
        "type": "object"
      },
      "PracticeConceptOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "label"
        ],
        "title": "PracticeConceptOut",
        "type": "object"
      },
      "PracticeQuestionOut": {
        "properties": {
          "answered": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PracticeAnsweredOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "choices": {
            "items": {
              "type": "string"
            },
            "title": "Choices",
            "type": "array"
          },
          "concept": {
            "$ref": "#/components/schemas/PracticeConceptOut"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "problem_number": {
            "title": "Problem Number",
            "type": "string"
          },
          "source_ref": {
            "title": "Source Ref",
            "type": "string"
          },
          "stem_md": {
            "title": "Stem Md",
            "type": "string"
          }
        },
        "required": [
          "id",
          "problem_number",
          "source_ref",
          "stem_md",
          "choices",
          "concept",
          "answered"
        ],
        "title": "PracticeQuestionOut",
        "type": "object"
      },
...
      "ProgressSummary": {
        "properties": {
          "scroll_pos": {
            "title": "Scroll Pos",
            "type": "number"
          },
          "section_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Id"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "section_id",
          "scroll_pos",
          "updated_at"
        ],
        "title": "ProgressSummary",
        "type": "object"
      },
...
      "SubmitPracticeAnswerIn": {
        "properties": {
          "selected_index": {
            "title": "Selected Index",
            "type": "integer"
          }
        },
        "required": [
          "selected_index"
        ],
        "title": "SubmitPracticeAnswerIn",
        "type": "object"
      },
      "SubmitPracticeAnswerOut": {
        "properties": {
          "already_answered": {
            "title": "Already Answered",
            "type": "boolean"
          },
          "concept": {
            "$ref": "#/components/schemas/PracticeConceptOut"
          },
          "correct": {
            "title": "Correct",
            "type": "boolean"
          },
          "correct_index": {
            "title": "Correct Index",
            "type": "integer"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "evidence_state": {
            "title": "Evidence State",
            "type": "string"
          },
          "explanation_md": {
            "title": "Explanation Md",
            "type": "string"
          },
          "question_id": {
            "title": "Question Id",
            "type": "string"
          },
          "readiness_estimate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readiness Estimate"
          },
          "selected_index": {
            "title": "Selected Index",
            "type": "integer"
          }
        },
        "required": [
          "question_id",
          "selected_index",
          "correct",
          "correct_index",
          "explanation_md",
          "concept",
          "readiness_estimate",
          "evidence_state",
          "evidence_count",
          "already_answered"
        ],
        "title": "SubmitPracticeAnswerOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SourceMind v2 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/courses/{course_id}": {
      "delete": {
        "operationId": "delete_course",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Course",
        "tags": [
          "courses"
        ]
      },
      "get": {
        "operationId": "get_course",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Course",
        "tags": [
          "courses"
        ]
      }
    },
...
    "/api/courses/{course_id}/diagnostics/validation/judgments": {
      "post": {
        "operationId": "submit_diagnostic_judgment",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticJudgmentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticJudgmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Diagnostic Judgment",
        "tags": [
          "diagnostic-validation"
        ]
      }
    },
    "/api/courses/{course_id}/diagnostics/validation/judgments/{judgment_id}/reason": {
      "patch": {
        "operationId": "record_diagnostic_disagreement_reason",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "judgment_id",
            "required": true,
            "schema": {
              "title": "Judgment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticDisagreementReasonIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticJudgmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Diagnostic Disagreement Reason",
        "tags": [
          "diagnostic-validation"
        ]
      }
    },
    "/api/courses/{course_id}/diagnostics/validation/next": {
      "get": {
        "operationId": "next_diagnostic_validation",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DiagnosticBlindCaseOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Next Diagnostic Validation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Next Diagnostic Validation",
        "tags": [
          "diagnostic-validation"
        ]
      }
    },
    "/api/courses/{course_id}/diagnostics/validation/summary": {
      "get": {
        "operationId": "diagnostic_validation_summary",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticValidationSummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diagnostic Validation Summary",
        "tags": [
          "diagnostic-validation"
        ]
      }
    },
...
    "/api/courses/{course_id}/practice-questions/{question_id}/answer": {
      "post": {
        "operationId": "submit_practice_answer",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitPracticeAnswerIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitPracticeAnswerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Practice Answer",
        "tags": [
          "practice"
        ]
      }
    },
...
    "/api/courses/{course_id}/sections/{section_id}/practice-assessment": {
      "get": {
        "operationId": "get_practice_assessment",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "section_id",
            "required": true,
            "schema": {
              "title": "Section Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeAssessmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Practice Assessment",
        "tags": [
          "practice"
        ]
      },
      "post": {
        "operationId": "start_practice_assessment",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "section_id",
            "required": true,
            "schema": {
              "title": "Section Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeAssessmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Practice Assessment",
        "tags": [
          "practice"
        ]
      }
    },
...
  }
}